Refactor ProductionPage to be more readable, extracting an EventCard and a ProductionMediaGallery
