+++ frontend/public/locales/en/translation.json
@@ -33,8 +33,7 @@
       "archive": "Archive",
       "defaultArtist": "viernulvier",
       "noDescription": "No additional description is available for this production.",
-      "dateTbd": "Date to be determined",
-      "locationTbd": "Location to be determined"
+      "noEvents": "No events recorded for this production."
     },
     "headerBrandAria": "VIERNULVIER Archive",
     "archiveLabel": "Archive",
