+++ frontend/public/locales/nl/translation.json
@@ -32,8 +32,7 @@
       "archive": "Archief",
       "defaultArtist": "viernulvier",
       "noDescription": "Geen extra beschrijving beschikbaar voor deze productie.",
-      "dateTbd": "Datum nog te bepalen",
-      "locationTbd": "Locatie nog te bepalen"
+      "noEvents": "Geen evenementen opgeslagen voor deze productie."
     },
     "headerBrandAria": "VIERNULVIER Archief",
     "archiveLabel": "Archief",
